Added image and buttons to Home.tsx, moved padding-left from body to sign-up-page class in App.scss Co-authored-by: Claudia Giancola <claudiaGiancola@users.noreply.github.com>
