+++ frontend/src/App.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { BrowserRouter, Routes, Route } from "react-router-dom"
 import Explore from "./pages/Explore"
-import Home from "./pages/Home"
+import Home from "./pages/Home/Home"
 import "./App.scss"
 import { CreateUser } from "./pages/CreateUser"
 import Header from "./Header/Header"
