+++ frontend/src/Header/Header.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent, render, screen } from "@testing-library/react"
 import { MemoryRouter, Route, Routes } from "react-router-dom";
 import Header from "../Header/Header";
 import { CreateUser } from "../pages/CreateUser";
-import Home from "../pages/Home";
+import Home from "../pages/Home/Home";
 import Explore from "../pages/Explore";
 
 describe("Header contains all the navigation items", () => {
