+++ frontend/src/Menu/Menu.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent, render, screen } from "@testing-library/react"
 import { MemoryRouter, Route, Routes } from "react-router-dom";
 import Hamburger from "../Hamburger/Hamburger";
 import { CreateUser } from "../pages/CreateUser";
-import Home from "../pages/Home";
+import Home from "../pages/Home/Home";
 import Explore from "../pages/Explore";
 
 describe("Menu toggles on/off when hamburger button clicked", () => {
