--- frontend/src/pages/Home.tsx
@@ -1,5 +1,0 @@
-const Home = () => {
-  return <h1>Home</h1>
-}
-
-export default Home
