+++ frontend/src/pages/Home/Home.scss
@@ -1,0 +1,4 @@
+#homepage-top-img {
+    margin: 0 0 0 0;
+    
+}
